refactor: 클래스 로직 변경으로 인한 스타일 수정
