+++ src/main/java/racingcar/InputCount.java
@@ -1,10 +1,10 @@
 package racingcar;
 
-import static racingcar.ConstantMessage.COUNT_MESSAGE;
-
 import camp.nextstep.edu.missionutils.Console;
 
 public class InputCount {
+
+    private static final String COUNT_MESSAGE = "시도할 횟수는 몇 회인가요?";
 
     public static String getInputCountData() {
         System.out.println(COUNT_MESSAGE);
